feat(multistream): match libopus for surround, projection and DTX edge cases
Port projection mixing in float32 with the tonality analysis on the
unmixed input, and fix CELT/Opus-layer differences the multistream
oracle exposed: OPUS_RESET_STATE clears the CELT energy mask,
transient_got_disabled is set on every frame whose transient flag does
not fit (silent frames advance consec_transient), VBR at
OPUS_BITRATE_MAX codes the whole budget, a CELT-only packet after a
hybrid one fades HB_gain back to unity, and the repacketizer pads with
zeros. The oracle gains a trace option for --ms-enc; the multistream
oracle covers 171 configurations and joins the CI gate.

+++ encoder_ms_budget.go
@@ -64,7 +64,7 @@ func (e *Encoder) surroundSILKRate(silkRate, bw int) int {
 		}
 	}
 	// Conservative rate reduction, we cut the masking in half.
-	maskingDepth := float32(maskSum/float32(end)) * float32(e.channels)
+	maskingDepth := float32(float32(maskSum/float32(end)) * float32(e.channels))
 	maskingDepth += 0.2
 	rateOffset := int(float32(float32(srate) * maskingDepth))
 	rateOffset = max(rateOffset, -2*silkRate/3)
